refactor(adapters): consolidate identical adapter methods into base class (#850 item 1)
Move identical and near-identical methods from the 6 standalone database
adapters into BaseDatabaseAdapter as overridable defaults, leaving only
genuinely engine-specific SQL generation in each subclass.

Shared defaults now in BaseDatabaseAdapter:
- buildWindowFunction (was byte-identical across all 6 engines)
- convertTimeDimensionResult, buildBooleanLiteral, convertFilterValue,
  prepareDateValue, isTimestampInteger, buildCaseWhen (SQLite overrides)
- buildConditionalAggregation: CASE WHEN default; Postgres/DuckDB override
  with the FILTER clause
- buildStringCondition: full switch delegating to caseInsensitiveLike() and
  regexCondition() hooks (ILIKE family by default; MySQL/SQLite/Databend
  override the case-insensitive hook, each engine overrides regex)
- buildAvg/buildStddev/buildVariance via a nullToZero() hook (COALESCE default;
  MySQL/SQLite override with IFNULL; SQLite returns null for stddev/variance;
  Databend overrides variance with the COVAR workaround)

Engine-specific methods (interval/date arithmetic, DATE_TRUNC vs DATE_FORMAT,
generate_series variants, capabilities, percentile, SQLite integer-timestamp
handling) remain per-adapter. Behavior is unchanged — verified against the full
multi-engine suite (Postgres/MySQL/SQLite all green).

Net ~770 lines removed. Also makes the unit-test SQL renderer recurse into
nested SQL fragments so COALESCE(AVG(x), 0)-style expressions render fully.
