+++ src/server/adapters/databend-adapter.ts
@@ -3,11 +3,15 @@
  * Implements Databend-specific SQL generation for time dimensions, string matching, and type casting
  * Databend uses pgcore (extends PgDialect) so is largely PostgreSQL-compatible
  * Key differences: no ILIKE, uses CASE WHEN for conditional aggregation, TIMESTAMPDIFF for time diffs
+ *
+ * Inherits shared defaults from BaseDatabaseAdapter (CASE WHEN conditional aggregation,
+ * COALESCE null-handling, standard window functions). Overrides the string-matching hooks
+ * (no native ILIKE) and VARIANCE (no native VAR_POP/VAR_SAMP).
  */
 
 import { sql, type SQL, type AnyColumn } from 'drizzle-orm'
 import type { TimeGranularity } from '../types'
-import { BaseDatabaseAdapter, type DatabaseCapabilities, type WindowFunctionType, type WindowFunctionConfig } from './base-adapter'
+import { BaseDatabaseAdapter, type DatabaseCapabilities } from './base-adapter'
 
 export class DatabendAdapter extends BaseDatabaseAdapter {
   getEngineType(): 'databend' {
@@ -68,22 +72,6 @@ export class DatabendAdapter extends BaseDatabaseAdapter {
   }
 
   /**
-   * Build Databend conditional aggregation using CASE WHEN
-   * FILTER clause support is uncertain in Databend, so use CASE WHEN for safety
-   */
-  buildConditionalAggregation(
-    aggFn: 'count' | 'avg' | 'min' | 'max' | 'sum',
-    expr: SQL | null,
-    condition: SQL
-  ): SQL {
-    const fnName = aggFn.toUpperCase()
-    if (aggFn === 'count' && !expr) {
-      return sql`${sql.raw(fnName)}(CASE WHEN ${condition} THEN 1 END)`
-    }
-    return sql`${sql.raw(fnName)}(CASE WHEN ${condition} THEN ${expr} END)`
-  }
-
-  /**
    * Build Databend date difference in periods using DATE_DIFF
    */
   buildDateDiffPeriods(startDate: SQL, endDate: SQL, unit: 'day' | 'week' | 'month'): SQL {
@@ -100,7 +88,7 @@ export class DatabendAdapter extends BaseDatabaseAdapter {
 
   /**
    * Build Databend time dimension using DATE_TRUNC function
-   * Databend supports DATE_TRUNC with quoted granularity like PostgreSQL
+   * Databend supports DATE_TRUNC with unquoted granularity keywords
    */
   buildTimeDimension(granularity: TimeGranularity, fieldExpr: AnyColumn | SQL): SQL {
     switch (granularity) {
@@ -126,32 +114,21 @@ export class DatabendAdapter extends BaseDatabaseAdapter {
   }
 
   /**
-   * Build Databend string matching conditions using LOWER+LIKE fallback
-   * Databend does not support ILIKE
+   * Databend has no ILIKE — use LOWER()+LIKE with SQL-side LOWER() on the pattern.
    */
-  buildStringCondition(fieldExpr: AnyColumn | SQL, operator: 'contains' | 'notContains' | 'startsWith' | 'endsWith' | 'like' | 'notLike' | 'ilike' | 'regex' | 'notRegex', value: string): SQL {
-    switch (operator) {
-      case 'contains':
-        return sql`LOWER(${fieldExpr}) LIKE LOWER(${`%${value}%`})`
-      case 'notContains':
-        return sql`LOWER(${fieldExpr}) NOT LIKE LOWER(${`%${value}%`})`
-      case 'startsWith':
-        return sql`LOWER(${fieldExpr}) LIKE LOWER(${`${value}%`})`
-      case 'endsWith':
-        return sql`LOWER(${fieldExpr}) LIKE LOWER(${`%${value}`})`
-      case 'like':
-        return sql`${fieldExpr} LIKE ${value}`
-      case 'notLike':
-        return sql`${fieldExpr} NOT LIKE ${value}`
-      case 'ilike':
-        return sql`LOWER(${fieldExpr}) LIKE LOWER(${value})`
-      case 'regex':
-        return sql`${fieldExpr} REGEXP ${value}`
-      case 'notRegex':
-        return sql`NOT (${fieldExpr} REGEXP ${value})`
-      default:
-        throw new Error(`Unsupported string operator: ${operator}`)
-    }
+  protected caseInsensitiveLike(fieldExpr: AnyColumn | SQL, pattern: string, negated: boolean): SQL {
+    return negated
+      ? sql`LOWER(${fieldExpr}) NOT LIKE LOWER(${pattern})`
+      : sql`LOWER(${fieldExpr}) LIKE LOWER(${pattern})`
+  }
+
+  /**
+   * Databend regex matching uses the REGEXP operator
+   */
+  protected regexCondition(fieldExpr: AnyColumn | SQL, value: string, negated: boolean): SQL {
+    return negated
+      ? sql`NOT (${fieldExpr} REGEXP ${value})`
+      : sql`${fieldExpr} REGEXP ${value}`
   }
 
   /**
@@ -169,62 +146,6 @@ export class DatabendAdapter extends BaseDatabaseAdapter {
       default:
         throw new Error(`Unsupported cast type: ${targetType}`)
     }
-  }
-
-  /**
-   * Build Databend AVG aggregation with COALESCE for NULL handling
-   */
-  buildAvg(fieldExpr: AnyColumn | SQL): SQL {
-    return sql`COALESCE(AVG(${fieldExpr}), 0)`
-  }
-
-  /**
-   * Build Databend CASE WHEN conditional expression
-   */
-  buildCaseWhen(conditions: Array<{ when: SQL; then: any }>, elseValue?: any): SQL {
-    const cases = conditions.map(c => sql`WHEN ${c.when} THEN ${c.then}`).reduce((acc, curr) => sql`${acc} ${curr}`)
-
-    if (elseValue !== undefined) {
-      return sql`CASE ${cases} ELSE ${elseValue} END`
-    }
-    return sql`CASE ${cases} END`
-  }
-
-  /**
-   * Build Databend boolean literal
-   * Databend uses TRUE/FALSE keywords
-   */
-  buildBooleanLiteral(value: boolean): SQL {
-    return value ? sql`TRUE` : sql`FALSE`
-  }
-
-  /**
-   * Convert filter values - Databend uses native types
-   */
-  convertFilterValue(value: any): any {
-    return value
-  }
-
-  /**
-   * Prepare date value for Databend
-   * Databend accepts Date objects directly
-   */
-  prepareDateValue(date: Date): any {
-    return date
-  }
-
-  /**
-   * Databend stores timestamps as native timestamp types
-   */
-  isTimestampInteger(): boolean {
-    return false
-  }
-
-  /**
-   * Databend time dimensions already return proper values
-   */
-  convertTimeDimensionResult(value: any): any {
-    return value
   }
 
   // ============================================
@@ -249,14 +170,6 @@ export class DatabendAdapter extends BaseDatabaseAdapter {
   }
 
   /**
-   * Build Databend STDDEV aggregation
-   */
-  buildStddev(fieldExpr: AnyColumn | SQL, useSample = false): SQL {
-    const fn = useSample ? 'STDDEV_SAMP' : 'STDDEV_POP'
-    return sql`COALESCE(${sql.raw(fn)}(${fieldExpr}), 0)`
-  }
-
-  /**
    * Build Databend VARIANCE aggregation
    * Databend doesn't have VAR_POP/VAR_SAMP, but COVAR_POP(x,x) = VAR_POP(x)
    * and COVAR_SAMP(x,x) = VAR_SAMP(x) mathematically
@@ -272,81 +185,5 @@ export class DatabendAdapter extends BaseDatabaseAdapter {
    */
   buildPercentile(_fieldExpr: AnyColumn | SQL, _percentile: number): SQL {
     throw new Error('Percentile functions are not yet supported for Databend')
-  }
-
-  /**
-   * Build Databend window function expression
-   * Databend has full window function support
-   */
-  buildWindowFunction(
-    type: WindowFunctionType,
-    fieldExpr: AnyColumn | SQL | null,
-    partitionBy?: (AnyColumn | SQL)[],
-    orderBy?: Array<{ field: AnyColumn | SQL; direction: 'asc' | 'desc' }>,
-    config?: WindowFunctionConfig
-  ): SQL {
-    // Build OVER clause components
-    const partitionClause = partitionBy && partitionBy.length > 0
-      ? sql`PARTITION BY ${sql.join(partitionBy, sql`, `)}`
-      : sql``
-
-    const orderClause = orderBy && orderBy.length > 0
-      ? sql`ORDER BY ${sql.join(orderBy.map(o =>
-          o.direction === 'desc' ? sql`${o.field} DESC` : sql`${o.field} ASC`
-        ), sql`, `)}`
-      : sql``
-
-    // Build frame clause if specified
-    let frameClause = sql``
-    if (config?.frame) {
-      const { type: frameType, start, end } = config.frame
-      const frameTypeStr = frameType.toUpperCase()
-
-      const startStr = start === 'unbounded' ? 'UNBOUNDED PRECEDING'
-        : typeof start === 'number' ? `${start} PRECEDING`
-        : 'CURRENT ROW'
-
-      const endStr = end === 'unbounded' ? 'UNBOUNDED FOLLOWING'
-        : end === 'current' ? 'CURRENT ROW'
-        : typeof end === 'number' ? `${end} FOLLOWING`
-        : 'CURRENT ROW'
-
-      frameClause = sql`${sql.raw(frameTypeStr)} BETWEEN ${sql.raw(startStr)} AND ${sql.raw(endStr)}`
-    }
-
-    // Combine OVER clause
-    const overParts: SQL[] = []
-    if (partitionBy && partitionBy.length > 0) overParts.push(partitionClause)
-    if (orderBy && orderBy.length > 0) overParts.push(orderClause)
-    if (config?.frame) overParts.push(frameClause)
-
-    const overContent = overParts.length > 0 ? sql.join(overParts, sql` `) : sql``
-    const over = sql`OVER (${overContent})`
-
-    // Build the window function based on type
-    switch (type) {
-      case 'lag':
-        return sql`LAG(${fieldExpr}, ${config?.offset ?? 1}${config?.defaultValue !== undefined ? sql`, ${config.defaultValue}` : sql``}) ${over}`
-      case 'lead':
-        return sql`LEAD(${fieldExpr}, ${config?.offset ?? 1}${config?.defaultValue !== undefined ? sql`, ${config.defaultValue}` : sql``}) ${over}`
-      case 'rank':
-        return sql`RANK() ${over}`
-      case 'denseRank':
-        return sql`DENSE_RANK() ${over}`
-      case 'rowNumber':
-        return sql`ROW_NUMBER() ${over}`
-      case 'ntile':
-        return sql`NTILE(${config?.nTile ?? 4}) ${over}`
-      case 'firstValue':
-        return sql`FIRST_VALUE(${fieldExpr}) ${over}`
-      case 'lastValue':
-        return sql`LAST_VALUE(${fieldExpr}) ${over}`
-      case 'movingAvg':
-        return sql`AVG(${fieldExpr}) ${over}`
-      case 'movingSum':
-        return sql`SUM(${fieldExpr}) ${over}`
-      default:
-        throw new Error(`Unsupported window function: ${type}`)
-    }
   }
 }
